processing_server/requirements.txt
aiohttp
beautifulsoup4==4.12.3
fastapi==0.112.0
langchain==0.2.12
langchain_community==0.2.11
langchain_openai==0.1.20
langchain_pinecone==0.1.3
pinecone==5.0.0
pydantic==2.8.2
pymongo==4.5.0
python-dotenv==1.0.1
Requests==2.32.3
uvicorn==0.30.5
pika==1.3.2